Hide tools not supported in 3d (see #1780)
This pull request is a part of #1780 and hides the following tools when the maptype is cesium:
 - Zoom In/Out
 - ScaleBox
 - Print
 - Locate
 - ShapeFile
 - Measure

It adds the display function as a allowed member of tools in a toolcontainer. If this fuction is present, it will be used to filter the tools to display. the entry. It is useful to hide entries in the ToolContainer other then hiding the window itself.

Also some documentation improved (e.g locate plugin, selectors, external types).