Fix #2558. Removed overwrite of OL event listener
`Map.jsx` overrides default event listeners but do not preserves function scopes.
This override is required only by the current implementation of DrawSupport that disables 'singleclick'
on map.
This pull request changes the semanthics of disableEventListeners to disable listeners of the map itself (not every click event) and for now I support only to disable 'singleclick'.
This way we can remove the override simplifying all the interaction and extending compatibility for every future update.