Min integration between lock, hide and collapse
Hide and lock have some issues with collapse functionality.
Rules applied to display/hide widgets are at plugin level. Instead they should be at global level,
to be shared between Widgets and WidgetsTrey plugins.

This minimal solution is hide to everyone (so even the user that can hide the widget) the widget in the tray.