fix #4079 units dropdown is not fully visible in measure tool
It turned out that the unit dropdown options in measure dialog gets
hidden behind the dialog and prevent user to select. This commit
fix and make the dropdown options visible and selectable.

Note:
This commits also change the border layout by allowing it to accept
style to override the default border inline style.