Issue #7504 Share map modal window not showing correctly on mobile
The fix applied for this issue consists of changing the component
that renders the content of the share modal window (from Dialog to ResizableModal).
The latter allows all content contained within the modal to be
completely responsive and scrollable so that the modal won't overflow
the mobile screen limits and the content can be seen scrolling up
when the device is in horizontal view mode.