Issue #7523 GeoCarousel improve highlighting of current item
Within the geostories functionality, when a new item (consisting of a point marker)
was added to a map within the geocarousel layout it was given a default
style. When the item was highlighted by the user, the default style was kept
and the selected item was added a chevron sign above the selected point,
besides when markers where too close to each other and overlapping the selected one was not
drawn on top to highlight the selection.
This made it difficult for users to see properly higlighted items.

The fix applied consisited of changing the style of the selected items, which colour is now red
and creating a layer to keep separated highlighted and non-highlighted items.
The highlighted items layer is added to the map after the highlighted one, so that
its z-index is higher and the highlighted items will always be drawn on top the
non-highlighted ones.