Issue #7504 Share map modal window not showing correctly on mobile
This commit includes changed to the CSS selector specifity
The selector logic is moved from the modal.less file
to the more specifc share.less file. The CSS change apply
padding to panel title and content so that it is not
pushed to the modal window boundaries, just for sharePanel.