Issue #7487 - Wrong string for group settings tooltip
This commit corrects the string for the tooltip on the
button to edit/confirm a name change on a layer or group
on the TOC.

This code change makes the component 'LayerNameEditField'
aware of the object users are interacting
with (either a group or a layer).
As a result new strings have got to be introduced in the
translations files.