MapStore2-C027 Issue #72 - Parsing layer name and metadate entries from csw
This commit implements and additional routine to grab values from
csw records, according to the new metadata specifications (RNDT / INSPIRE)
the layer name is no longer specified explicitly as an attribute in the
<dc> entries and has got to be sought on the extended URI for the wms
layer.

The function parses the dc entries, finding the one that contains
the layer wms url and extracts from the querystring
the values that are needed by MapStoreto correctly add
the wms layer to the map from the csw record information.