MapStore2-C027 Issue 72 - Parsing layer name and metadate entries from csw
This commit completes the dev work to correctly parse csw records,
according to the new metadata specifications (RNDT / INSPIRE)
the attributes and the structure of <dc:URI> has changed and
no longer contains the attribute name, said attribute was
used to display the titles in the respective <a> in MapStore
to download locally the layer resource examples of formats
are shp, kml.

There is a new routine that parses the dc entries
and formats the resulting <a> tags sourcing it
from the csw resource title. Such piece of information
is found in the csw record, in the <dc:title>