MapStore2-C027 Issue 72 - Parsing layer name and metadate entries from csw
When testing on C027 backend there are cases where in the returned XML
csw response protocol is not returned. The resulting XML entry contains only
value (as follows <dc:URI>http://datigis.comune.intranet/geoserver/wms?tiled=true&amp;version=1.1.1</dc:URI>)
Though this might not be a correct entry acoording to the specification, we implemented
code that makes up for the lack of protocol inferring it from the only available
data, the value.