C027 Issue 72 - layer name parsing (#7575)
* MapStore2-C027 Issue #72 - Parsing layer name and metadate entries from csw

This commit implements and additional routine to grab values from
csw records, according to the new metadata specifications (RNDT / INSPIRE)
the layer name is no longer specified explicitly as an attribute in the
<dc> entries and has got to be sought on the extended URI for the wms
layer.

The function parses the dc entries, finding the one that contains
the layer wms url and extracts from the querystring
the values that are needed by MapStoreto correctly add
the wms layer to the map from the csw record information.

* MapStore2-C027 Issue 72 - Parsing layer name and metadate entries from csw

Second commit because I forgot to stage the latest changes before.

This commit implements and additional routine to grab values from
csw records, according to the new metadata specifications (RNDT / INSPIRE)
the layer name is no longer specified explicitly as an attribute in the
<dc> entries and has got to be sought on the extended URI for the wms
layer.

The function parses the dc entries, finding the one that contains
the layer wms url and extracts from the querystring
the values that are needed by MapStoreto correctly add
the wms layer to the map from the csw record information.

* MapStore2-C027 Issue 72 - Parsing layer name and metadate entries from csw

This commit completes the dev work to correctly parse csw records,
according to the new metadata specifications (RNDT / INSPIRE)
the attributes and the structure of <dc:URI> has changed and
no longer contains the attribute name, said attribute was
used to display the titles in the respective <a> in MapStore
to download locally the layer resource examples of formats
are shp, kml.

There is a new routine that parses the dc entries
and formats the resulting <a> tags sourcing it
from the csw resource title. Such piece of information
is found in the csw record, in the <dc:title>

* MapStore2-C027 Issue 72 - Parsing layer name and metadate entries from csw

When testing on C027 backend there are cases where in the returned XML
csw response protocol is not returned. The resulting XML entry contains only
value (as follows <dc:URI>http://datigis.comune.intranet/geoserver/wms?tiled=true&amp;version=1.1.1</dc:URI>)
Though this might not be a correct entry acoording to the specification, we implemented
code that makes up for the lack of protocol inferring it from the only available
data, the value.

* MapStore2-C027 Issue 72 - Parsing layer name and metadata entries from csw

This commit contains the changes applied after having included suggestions
made within the code review.

* MapStore2-C027 Issue 72 - Parsing layer name and metadata entries from csw

Fixed lint errors

* MapStore2-C027 Issue geosolutions-it#72 - Parsing layer name and metadata entries from csw

changed tests target for csw record holding WMS, including protocol key.

* MapStore2-C027 Issue 72 - Parsing layer name and metadata entries from csw

This commit redesigns the method to weed out csw dc entries that won't hold
any wms information. As we now remanage the structure of the object according
to RNDT specs we have to move the filter logic after the entries holding the
WMS endpoints have been extracted

Co-authored-by: Alessandro Cristofori <alessandro.cristofori@geosolutionsgroup.com>