Issue #7504 Share map modal window not showing correctly on mobile
This commit contains changes to the CSS rules to make the share panel
modal window height smaller when the screen width is larger than 768px
(bootstrap sm breakpoint). Also, when the modal is made draggable,
for example when the share button is not pressed on the home page,
the modal window pops up in a higher position on the screen so that
the modal won't 'fall' to bottom when showing in smaller screen sizes.