Issue #7594 Define colour-coded classification for charts widgets
This first commit implements the backbone of the workflow
that will allow users to classify charts values other than
with data defined on the x and y axes.

Changes to the code in this commit include the creation of
a modal window that will pop up when the user defines the classes.
Within the modal there is a dropdown to define which attribute
the user wants to style the chart with and an initial implementation
of a colour picker to choose the colour for the selected values of the attribute.