#7246 Do not export styles when map is exported
Export to MapStore format will not contain "availableStyles" array regardless if it exists in current map state or not.
Export to WMC format will not call capabilities load, styles will not be included in exported map.

Both cases do not affect how layers are styled when map is being re-imported.