add serverType option to wms & csw catalog entries #8167
if undefined or set to "geoserver", the behaviour is the unchanged. If set to
"no-vendor", then TILED query parameter isnt sen't to the remote wms server, and
the "tiled" and "localizedLayerStyles" options are hidden in the layer
parameters.

more vendor options can be sent/avoided this way. Improves interoperability
when displaying and printing layers coming from mapproxy.

closes #7809, #7877 & #7879