Make Validator a wrapper for predicate
In reality all a validator does is expose two functions that filter
based on a format-dependent predicate/criterium.

The only exception being String format, where both valid and invalid
responses require need to be of type string, but this is likely a
mistake.

Instead of constructing them one by one we extract the criteria into the
Predicate object and make the Validator a wrapper containing the
previous array-based functions as well as the unterlying predicate
itself.

This way, the criterium is very clear and it is guaranteed that any
response that is not valid is indeed invalid.

It also makes it easier to expose the predicate for direct use so that
one does not need to cast single responses to arrays and then check
whether the remain in the filtered array.