Fix test expectation
There was something weird happening in the previous code where some TEXT
responses were neither valid nor invalid in the sense that it would be
filtered out both by 'getValidResponses' as well as
'getNoValidResponses'. The only responses that should occur in neither
of the arrays are 'falsy' responses like 'undefined'.

This lead to a response with format 'TEXT' and content not of type
'string' to be neither valid nor invalid. I changed this when I
introduced predicates and defining anything that is not valid to
be invalid. Thus, none of the responses in the example are valid TEXT
responses but 'undefined' is filtered. So there should be 3 elements.