Bugfix: Link for shared maps without context
If you use the share functionality in a map, that is not created out of a context,
the shared link will not show the correct center and zoom, as long as the /config
part in the URL is missing.
As the documentations states in
https://docs.mapstore.geosolutionsgroup.com/en/latest/developer-guide/map-query-parameters/#center-zoom
and in
https://docs.mapstore.geosolutionsgroup.com/en/latest/developer-guide/map-query-parameters/#marker-zoom
that such a /config part should be part of configured URLs, this fix adds the /config part

Test is adapted to the new solution.

Bugfix on behalf of DB Systel GmbH